Rename dir and fix gitignore
