+++ .gitignore
@@ -116,15 +116,13 @@ output.ipynb
 # don't save any data
 classification/data/*
 data/
+!contrib/action_recognition/r2p1d/**
 
 # don't save .swp files
 *.swp
 
 # don't save .csv files
 *.csv
-
-# don't save data dir
-data
 
 # don't save pickles
 *.pkl
