+++ contrib/action_recognition/r2p1d/data/samples/label.txt
@@ -1,0 +1,1 @@
+drinking 100
