+++ contrib/action_recognition/r2p1d/vu/utils/__init__.py
@@ -1,0 +1,1 @@
+from .common import Config, system_info
