--- contrib/action_recognition/r2plus1d/vu/utils/__init__.py
@@ -1,1 +1,0 @@
-from .common import Config, system_info
